Moved conky file.
Moved the Conky file to the .i3 folder.
The Conky config contains some conditional colour settings. For example when
your RAM is sky high or when your available disk space is running low. Also it
will display the number of new mails, but only if there are any. Also this is
limited to the number of mails in my inbox.