Removed short_text
Removed short_text, because it is pretty useless. It didn't work like I thought
it would. Didn't read all the documentation :D :X
Thought it would be used if the i3bar as a whole got too big, but it's just
about the single attribute. So short_text should got together with min_width.